Splits tests into several files.
Closes #9
